- TimerManager (ultra coooool ! look weapon code to see a example!) - fix bug reload and shoot at the same time

+++ src/Game/Troma/Troma/WeaponInfo.cs
@@ -14,7 +14,7 @@ namespace Troma
         public int Loader;
 
         public bool Automatic;
-        public float ROF;
+        public uint ROF; // en ms
 
         public string Model;
